% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{fbEval}
\alias{fbEval}
\title{Apply frequency band model evaluation}
\usage{
fbEval(dat, id = "monid", truth = "monitor", pred = "FAQSD")
}
\arguments{
\item{dat}{Dataset with columns of date, id (monid), truth (e.g., monitoring data), and pred (e.g., model predictions).  All other columns must refer to spatial location/model}

\item{id}{Name of id column (e.g., monitor ID)}

\item{truth}{Name of column of true concentrations}

\item{pred}{Name of column of model predictions}
}
\description{
Apply frequency band model evaluation
}
\examples{
 
data(PM25)
fbresult <- fbEval(pm)
fbresult$meval
}
\author{
Jenna Krall
}
