% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{fbTimeSeries}
\alias{fbTimeSeries}
\title{Frequency band time series}
\usage{
fbTimeSeries(dat, id = "monid")
}
\arguments{
\item{dat}{Dataset with columns of date, id (monid), and value.  All other columns must refer to spatial location/model}

\item{id}{Name of id column (e.g., monitor ID)}
}
\description{
Frequency band time series
}
\author{
Jenna Krall, Joshua P. Keller
}
