% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{lvr}
\alias{lvr}
\title{\code{lvr} Log variance ratio}
\usage{
lvr(truth, error)
}
\arguments{
\item{truth}{True time series}

\item{error}{Error-prone (predicted) time series}
}
\description{
\code{lvr} Log variance ratio
}
