% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM25.R
\docType{data}
\name{PM25}
\alias{PM25}
\title{PM2.5 monitoring data and FAQSD model predictions}
\format{
A data frame with 19888 rows and 8 variables
}
\description{
A dataset of PM2.5 monitoring data and predictions from fused CMAQ for 17 monitors.
}
\details{
\itemize{
\item date. Date of observation
\item monid. Monitor ID
\item fips. FIPS code
\item city. Major urban area where monitor is located
\item longitude. Longitude
\item latitude. Latitude
\item FAQSD.  Predictions from the FAQSD model
\item monitor. Monitor concentrations
}
}
\keyword{datasets}
