% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{rmsefun}
\alias{rmsefun}
\title{\code{rmsefun} Root mean square error}
\usage{
rmsefun(truth, error, sd = F)
}
\arguments{
\item{truth}{True time series}

\item{error}{Error-prone (predicted) time series}

\item{sd}{Whether to standardize by true standard deviation}
}
\description{
\code{rmsefun} Root mean square error
}
