% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-functions.R
\name{wrap_tsdecomp_dat}
\alias{wrap_tsdecomp_dat}
\title{Wrapper function to run tsdecomp for PM2.5 data}
\usage{
wrap_tsdecomp_dat(dat, decomp_breaks = NULL, season = T)
}
\arguments{
\item{dat}{Dataset with column named value indicating time series}

\item{decomp_breaks}{Vector of breaks for decomposition (if null, specify to see seasonal, monthly, acute in simulated data)}

\item{season}{Whether to return season, month, acute (as in Krall et al. 2021+)}
}
\description{
Wrapper function to run tsdecomp for PM2.5 data
}
\author{
Jenna Krall, Joshua P. Keller
}
